FullContract now ignores history and generates complete compute lists.
The checking which was formerly in FullContract must be moved to
equation computer.

The dependents of a CTP must now be specified in a member of that CTP.
All dependencies of a CTP are now specified in it's entry in A_compute map.

Must add skip to CTP to prevent repetion of contractions performed on CMTP

This all needs thorough checking