Added in new routines for single tensor contraction, not merged.
Mainly commited for reference: assert statement was tripping- saying
the RMAWindow wasn't initialized. Replacing the routine with a
previous version has fixed this, but I am not clear why.

So if that happens again, diff with this commit. If there is no
clear cause, then there is probably a nasty bug.