Both tensor contraction routines appear to work with blocked sparse tensors.
Fixed merging ranges bug.
Added add_block with vector argument to tensor.

This has been tested for a few different cases, and seems to hold up.
So this should be the REFERENCE VERSION if things go wrong later,
even though it is very slow at the moment due to the
excessive amount of printing.