Fixed bug in Full Contract.
A compute lists for intermediate tensors were
not begin put into the compute_map in binary contract different
tensors diff tensors. That it ran through with this bug is strange.

Something is inconsistent; results hop between a few definite values
on rerunning. Values must be set somewhere, as the definite values
are very consistent, even after changing the code and recompiling.