New same tensor contraction routine is working.
Problem in gammas is resolved.

Problem with old contraction routine is also fixed,
but the new one is probably better for handling
excitation operators.

Consider if reordering in gamma can be altered; routine
is much faster if ctrs are already at the front.

one of the gamma_4idx test is giving a false negative,  an
alternative to this test must be implemented