Reverted to dvec routines for sigmas, split out reworked eqn_comp_gammas
This is not remotely efficient, but it does seem to work provided Bagel
formatting is used.

Sigma routines work correctly with tensor formatting. So does first gamma, but
above that exactly the same problem as before appears. There
is a clearly something issue in the tensor-vector contraction routines,
but this might not be all.