New routine to automate generation of expression and BraKet lists etc. from op_name vector
Found issue when tensors decomposed automatically contraction is
the parts are not found. New contraction routine, or at least
interface is needed.

Now has factors for individual BraKets, but these will be lost
when merged. This must be fixed. Similarly, expression name
generation is now automated, but this does not include factors.