Adding Tensop_Prep class.
The motivation behind this is that several of the members of
tensop are common to all states and will never change, and so
should always be const. However, the sparsity, and hence what
is fed to gamma_generator and the resulting compute list,
is state dependent, and may vary as the calculation progresses.

So first we initialize a tensop_prep object at the beginning of
the calculation, and use this to build the various TensOp objects.
These TensOps have a pointer to the TensOp_prep, and only differ
from one another by the range maps they include, which can
be updated over the calculation to reflect the sparsity of the
relevant tensors.