Ammendment : <M|E_{Omega}*A|N> Results agree up to factor of 2, and anomalous minus sign in vvaa term.
Here A has the ranges of H, but I have set is so A is a generated
tensor which has no symmetry.

It is hard to tell how serious the issue with the minus sign is;
it could be a typo somewhere, or it could be a fundamental flaw,
probably in gamma_generator.