Seperating out functionality into new computer class.
Ultimately, proptool should translate everything into tensors/indexranges,
and feed this to System_Computer. System_Computer should be blind to all
Bagel classes (bar SMITH::Tensor_ and SMITH::IndexRange_).

This is why the MOInt_Computer (and not just the MOInt_Init) is
 constructed in proptool; computer classes should
be constructed within System_Computer, but the mointegrals
make such extensive use of Bagel Types this would be impractical.

At the moment, System_Computer takes the entirety of Smith_Info, I
am not sure if this is for the best, but it is simplest, so keep it for now.