Fixing issue in contractions involving 3+ tensors and direct product
Now altered class structure of tensops and made various other changes elsewhere
to accomodate this.

This is to avoid problmes with the nesting when it comes to reordering
and doing direct products of the blocks.

The issue with shared from this is technically fixed, in so much
as it is safe, but I think there should be a better way to
deal with the sub_tensops  and ctp_vecs