start of documentation of program,
This is not the place where it should ultimately be kept, and
is not in the correct format, but thought here might be convenient
for the time being.

This is a very rough draft, and is only the introduction at present.
I will extend it to cover the different modules of the program
(focusing on the algebraic manipulator).