More pruning and basic alterations to GGR constructor.
Now standardized ordering is forcing it to crash for flipped
terms (as it should; is now fully reordering so as to minimize
terms). Now must alter ctp map build to only keep the standard
orders. Keep range_blocks in weird order specific to a BraKet's MTens;
multiple MTens and rangeblocks with shuffled orders and factors from symmetry,
but task list should always be written in terms of the same set of CTPs.