trying and failing to add in range transformations.
this should be trivial, but is not due to my moronic attempt
at const correctness in the tensop classes.

If this difficulty persists, serious consideration must be given
to reverting the tensop classes to their original form; this is not the
first day I have wasted on something so stupid.