Reworking BraKet/AContrib/states/civec/gamma etc.
Different kinds of expressions and terms should be controlled
by class, rather than with members called "type" and ifs.

Also for factor control; have one factor as pairs of doubles,
another, different pair for real and Im factors (to track transformations).
Templating is kept as re and im will be split so task list generation
can be selected to keep to real arithmetic, and checking sparsity.

Junked much of a_contrib_info; it was constructed on the basis of the
old version of the gamma generator, and was far too complicated.