Changing blocks/op_info construction.
Fixed earlier problem with incorrect/excessive block creation fetching.

Now have strange recursion in definition of canonical definitions which may cause
problems, but seems OK. The op_info and range block need to be cleaned.

Need to sort out handling of state symmetry in Op_info so that it works
within the new canonical members formalism.

Must include transformation of aops when building canonical variant;
not an issue here, but could be relevant in special cases.