Fixed various things in tensor arithmetic.
Main changes in add_tensor_along_trace. Should now be OK when blocks
are split and uneven, when summand is reordered, and when
indexes on target are not adjacent. Reordering of summand is
never done explicitly, but merged into second (inverse) reordering of
target. Needs testing with a greater number of tensors to be safe.
Also needs cleaning; using [] at present for clarity.