Reworking the gamma reorderer.
This is important blockwise testing, but I do  not think it should have impacted
the final results (and it hasn't in the tests which I've done).

Fixed bug in proj_onto_map (incorrect use of ggen member, not argument; should rethinknaming),
this would not have impacted the current tests, and hasn't, but could have shown up later

Also added in raw version of gamma_intermediates; these should now be incorporated. Again,
this is to sort out the anomalous behaviour when testing the swapping algorithms; whilst in
the test cases I can parse I see nothing wrong, but it is possible the excessive use of shared pointers
is causing some errors in the more complicated terms.